# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "studybot api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        },
        {
            "url": "https://api.studybot.it",
            "description": "Live server"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/auth/linkDigreg": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Link a digreg account to a discord account",
                "description": "Link a digreg account to a discord account",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "digregToken": {
                                        "type": "string",
                                        "required": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid digregToken"
                    },
                    "401": {
                        "description": "Not logged in"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/auth/unlinkDigreg": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Unlink a digreg account from a discord account",
                "description": "Unlink a digreg account from a discord account",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Not logged in"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}